Fixed up the File and FileSystem classes to be 64-bit clean by using off_t instead of int and long for file offsets.

+++ src/common/OW_Enumeration.hpp
@@ -180,12 +180,12 @@ private:
 				"signature");
 		}
 		
-		UInt32 whence=0;
+		off_t whence=0;
 		if ((whence = f.seek(-sizeof(size), SEEK_END)) == -1)
 		{
 			OW_THROW(EnumerationException, "Failure to seek");
 		}
-		if(f.read(reinterpret_cast<char*>(&size), sizeof(size), (long)whence) != sizeof(size))
+		if(f.read(reinterpret_cast<char*>(&size), sizeof(size), whence) != sizeof(size))
 		{
 			OW_THROW(EnumerationException, "Failure to read enumeration "
 				"size");
